controller: add proxy message
Add a new string message for sending proxy details.
CONTROLLER_PROXY (ControllerData, string)

The value must be of the form [protocol://]<host>[:port]

The proxy string may be specified with a protocol:// prefix to specify
alternative proxy protocols.

If no protocol is specified in the proxy string or if the string
doesn't match a supported one, the proxy will be treated as a HTTP
proxy.