vd_agent: Add a VD_AGENT_FILE_XFER_STATUS_SUCCESS result value
Currently the sender of a file xfer assumes success on having send the
last data bytes. But the transfer may still fail on the other side.

This commits adds a VD_AGENT_FILE_XFER_STATUS_SUCCESS result value instead.

Since we have not done an agent release with file-xfer support yet, the client
code can simply assume that it will always get *a* VD_AGENT_FILE_XFER_STATUS_
message for a transfer now. As for the existing spice-gtk release with
file-xfer support, if it talks to an agent sending this message, this will
trigger a g_return_if_fail, which is not really pretty, but has no negative
side-effects other then an error message being logged.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>