Add unix GL scanout messages
Add 2 new messages to the display channel to stream pre-rendered GL
images of the display. This is only possible when the client supports
SPICE_DISPLAY_CAP_GL_SCANOUT capability.

The first message, SPICE_MSG_DISPLAY_GL_SCANOUT_UNIX, sends a gl image
file handle via socket ancillary data, and can be imported in a GL
context with the help of eglCreateImageKHR(). It can then be displayed
with the given display coordinate (x/y/w/h) within the image.

When the image is updated, either a new image replaces the current image
or a SPICE_MSG_DISPLAY_GL_SCANOUT_UPDATE message is sent with the
coordinate of the updated region.

The relation with the existing display channel messages is that all
other messages are unchanged, the last drawing command received must be
displayed, however the scanout display is all or nothing. Consequently,
if a draw is received, the display must be switched to the drawn
display. In other words, if the last message received is a GL scanout
the display should switch to the GL scanout image, if it's a draw
message the display should be switched to the client rendered canvas.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>